requirements.txt
python-dotenv
requests
apify_client
supabase